DM-4483:more editing
