+++ src/firefly/js/drawingLayers/WebGrid.js
@@ -12,13 +12,31 @@ import {primePlot} from '../visualize/PlotViewUtil.js';
 import {makeFactoryDef} from '../visualize/draw/DrawLayerFactory.js';
 import {getUIComponent} from './WebGridUI.jsx';
 import { getData } from './ComputeWebGridData.js';
- import DrawLayerCntlr, {dispatchModifyCustomField, dispatchForceDrawLayerUpdate} from '../visualize/DrawLayerCntlr.js';
+import DrawLayerCntlr, {dispatchModifyCustomField, dispatchForceDrawLayerUpdate} from '../visualize/DrawLayerCntlr.js';
+import AppDataCntlr from '../core/AppDataCntlr.js';
+ import ShapeDataObj from '../visualize/draw/ShapeDataObj.js';
+ import CoordinateSys from '../visualize/CoordSys.js';
+ import CoordUtil from '../visualize/CoordUtil.js';
+ import {get} from 'lodash';
 
 
+ export const COORDIANTE_PREFERENCE = 'coordinate';
+
+const  coordinateArray = [
+     {coordName:'eq2000hms',        csys:CoordinateSys.EQ_J2000},
+     {coordName:'eq2000dcm',        csys:CoordinateSys.EQ_J2000},
+     {coordName:'eqb1950hms',       csys:CoordinateSys.EQ_B1950},
+     {coordName:'eqb1950dcm',       csys: CoordinateSys.EQ_B1950},
+     {coordName:'galactic',         csys:CoordinateSys.GALACTIC},
+     {coordName:'superGalactic',    csys:CoordinateSys.SUPERGALACTIC},
+     {coordName:'epj2000',          csys:CoordinateSys.ECL_J2000},
+     {coordName:'epb1950',          csys:CoordinateSys.ECL_B1950}
+
+ ];
 
 const ID= 'WEB_GRID';
 const TYPE_ID= 'WEB_GRID_TYPE';
-const factoryDef= makeFactoryDef(TYPE_ID,creator, getDrawData,null, null,getUIComponent);
+const factoryDef= makeFactoryDef(TYPE_ID,creator, getDrawData,getLayerChanges, null,getUIComponent);
 export default {factoryDef, TYPE_ID}; // every draw layer must default export with factoryDef and TYPE_ID
 /**
  * Color used to draw the grid
@@ -63,8 +81,7 @@ function getDrawData(dataType, plotId, drawLayer, action, lastDataRet){
      var cc= CsysConverter.make(plot);
      if (!cc) return null;
 
-     var drawDataArray= getData(plot, cc);
-     return drawDataArray || lastDataRet ;
+     return lastDataRet ||getData(plot, cc) ;
  }
 
 
@@ -77,3 +94,41 @@ function getDrawData(dataType, plotId, drawLayer, action, lastDataRet){
      return Object.assign({coordinate:{drawData:dd}});     // return the drawLayer changes
  }
 
+ /**
+  * This method prepare the input parameters needed for calculating the drawing data array
+  * @param plot
+  * @returns {{width: (dataWidth|*), height: (*|dataHeight), screenWidth: *, csys: *, labelFormat: string}}
+  */
+ export function getDrawLayerParameters(plot){
+     var width = plot.dataWidth;
+     var height = plot.dataHeight;
+     var screenWidth = plot.screenSize.width;
+
+     var csysName = AppDataCntlr.getPreference(COORDIANTE_PREFERENCE);
+     if (!csysName) {
+         csysName='eq2000hms';//set default
+     }
+     var csys=getCoordinateSystem(csysName);
+
+
+     var labelFormat=csysName.endsWith('hms')? 'hms':'dcm';
+
+     return {width, height, screenWidth, csys,labelFormat};
+ }
+
+ /**
+  * Return a Coordinate object for a given coordinate name
+  * @param csysName
+  * @returns {*}
+  */
+
+ function getCoordinateSystem(csysName) {
+
+     for (let i = 0; i < coordinateArray.length; i += 1) {
+         if (get(coordinateArray[i], 'coordName') === csysName) {
+             return coordinateArray[i].csys;
+
+         }
+     }
+
+ }
